Update of Technical Charter & Governance
The technical charter and the governance have been revised on 15 October 2021 and have been subject to approval by LF Projects.
All the TSC maintainers are invited to vote on the update of the technical charter and the governance: agree vs. disagree.
Maintainers who have not voted by 23:59 November 7, 2021 will be counted as approval votes by default.

The two documents will become effective when 2/3 of TSC members vote for approval.